refactor(paint-boolean-ops): split helpers and hotfixes into separate modules
- Extract pure paper.js utilities into paper-helpers.js
- Extract paper.js prototype patches into hotfixes.js
- Fix separator between Back/Front and shaping section disappearing
  after resize cycle: re-apply dashed-border class to frontBackRow on
  every waitForElement iteration, not just the first
